==> Scripts/Differential-Expression/04_ENRICHR.ipynb <==
!pip install 'folium==0.9.1'
!pip install "gseapy==0.10.7"

## IMPORT STATEMENTS

import pandas as pd
import gseapy as gp
import numpy as np

## LOAD FILE WITH GENE LISTS

## MAKES DICT WITH ALL SHEETS (NAME : DF)
path_to_list = "/content/drive/MyDrive/2023/ENRICHR-GENE-LISTS.xlsx"
my_gene_lists = pd.read_excel(path_to_list, sheet_name=None,header=None)
#path_to_background = "/content/drive/MyDrive/2023/ALL_GENES.xlsx"
#my_background = pd.read_excel(path_to_background, header=None)
## ADDING BACKGROUND DIDN'T SEEM TO MAKE A DIFFERENCE


## Run EnrichR and save results

writer_path = "/content/drive/MyDrive/2023/ENRICHR/ENRICHR-RESULTS-PVAL-AND-OD-FILTERED.xlsx"
with pd.ExcelWriter(writer_path) as writer:
  for key,value in my_gene_lists.items():
    enr = gp.enrichr(gene_list=value.dropna(), ## drop no hgnc symbol blanks
                    gene_sets="MSigDB_Hallmark_2020",
                    outdir = "/content/drive/MyDrive/2023/ENRICHR/" + key + "-unfiltered",
                    cutoff=0.25)
    enr.results["Odds Ratio"] = enr.results["Odds Ratio"].round(1) ## ROUND ODDS RATIO
    my_filter = enr.results.loc[(enr.results["Adjusted P-value"] < 0.1) & (enr.results["Odds Ratio"].round(1) >= 3)]
    my_filter.to_excel(writer,sheet_name = key,index=False,columns = ["Term","Overlap","P-value","Adjusted P-value","Odds Ratio","Genes"])
